ScryfallClient: Update API to return an Optional<T> for all endpoints
This patch replaces the default return value for each API call with an Optional of the same
type. This allows errors to be propagated back to the caller.

I am testing the feel of the API using an 'ErrorOr' approach, though Optional is not really
the correct vehicle for this.